Address Cursor comments re hotkeys, slow transition issues
