+++ docs/app/(home)/layout.tsx
@@ -1,25 +1,12 @@
 import { HomeLayout } from 'fumadocs-ui/layouts/home';
-import { headers } from 'next/headers';
 import { baseOptions } from '@/lib/layout.shared';
 import { EveChatMount } from '@/components/eve-chat-mount';
-import { DocsProductProvider } from '@/components/docs-product-context';
-import {
-  DEFAULT_DOCS_PRODUCT,
-  DOCS_PRODUCT_HEADER,
-  parseDocsProduct,
-} from '@/lib/home-navigation';
 
-export default async function Layout({ children }: LayoutProps<'/'>) {
-  const requestHeaders = await headers();
-  const initialProduct =
-    parseDocsProduct(requestHeaders.get(DOCS_PRODUCT_HEADER)) ?? DEFAULT_DOCS_PRODUCT;
-
+export default function Layout({ children }: LayoutProps<'/'>) {
   return (
-    <DocsProductProvider initialProduct={initialProduct}>
-      <HomeLayout {...baseOptions()}>
-        {children}
-        <EveChatMount />
-      </HomeLayout>
-    </DocsProductProvider>
+    <HomeLayout {...baseOptions()}>
+      {children}
+      <EveChatMount />
+    </HomeLayout>
   );
 }
